Refactor react client/server code to promote single use functions/components
Splits out various parts of the the server/client render code into more
granular functions which are "more" single use than before. Enables
users of framework to gradually transition to running these functions
directly instead of relying on `Client({...})` and `Server({...})`,
enabling apps to decide how and when to use the functions. Useful as
a stepping stone to enable apps to implement `getConfigureStore` themselves
and therefore be able to implement hot reloading of reducers